+++ src/Rocketeer/Facades/Rocketeer.php
@@ -8,7 +8,6 @@ use Illuminate\Support\Facades\Facade;
  */
 class Rocketeer extends Facade
 {
-
 	/**
 	 * Get the registered name of the component.
 	 *