+++ src/Rocketeer/Scm/ScmInterface.php
@@ -4,8 +4,14 @@ namespace Rocketeer\Scm;
 /**
  * The interface for all SCMs
  */
-interface Scm
+interface ScmInterface
 {
+	/**
+	 * Check if the SCM is available
+	 *
+	 * @return string
+	 */
+	public function check();
 
 	/**
 	 * Get the current state