+++ src/Rocketeer/Tasks/Deploy.php
@@ -1,7 +1,7 @@
 <?php
 namespace Rocketeer\Tasks;
 
-use Rocketeer\Tasks\Abstracts\Task;
+use Rocketeer\Traits\Task;
 
 /**
  * Deploy the website