+++ src/Rocketeer/Tasks/Rollback.php
@@ -1,7 +1,7 @@
 <?php
 namespace Rocketeer\Tasks;
 
-use Rocketeer\Tasks\Abstracts\Task;
+use Rocketeer\Traits\Task;
 
 /**
  * Rollback to the previous release, or to a specific one