+++ src/Rocketeer/Traits/Task.php
@@ -1,5 +1,5 @@
 <?php
-namespace Rocketeer\Tasks\Abstracts;
+namespace Rocketeer\Traits;
 
 use Rocketeer\Bash;
 
@@ -238,7 +238,7 @@ abstract class Task extends Bash
 	{
 		$composer = $this->which('composer');
 		if (!$composer and file_exists($this->app['path.base'].'/composer.phar')) {
-			$composer = 'composer.phar';
+			$composer = 'php composer.phar';
 		}
 
 		return $composer;