+++ tests/TasksQueueTest.php
@@ -7,7 +7,14 @@ class TasksQueueTest extends RocketeerTests
 	{
 		$task = $this->tasksQueue()->buildTask('Rocketeer\Tasks\Deploy');
 
-		$this->assertInstanceOf('Rocketeer\Tasks\Abstracts\Task', $task);
+		$this->assertInstanceOf('Rocketeer\Traits\Task', $task);
+	}
+
+	public function testCanAddCommandsToArtisan()
+	{
+		$command = $this->tasksQueue()->add('Rocketeer\Tasks\Deploy');
+		$this->assertInstanceOf('Rocketeer\Commands\BaseTaskCommand', $command);
+		$this->assertInstanceOf('Rocketeer\Tasks\Deploy', $command->getTask());
 	}
 
 	public function testCanGetTasksBeforeOrAfterAnotherTask()